+++ RELEASE.md
@@ -1,0 +1,4 @@
+Release type: patch
+
+This release fixes [issue #96](https://github.com/nrbnlulu/qtgql/issues/96).
+
