+++ docs/helpers/itemsystem.md
@@ -12,7 +12,7 @@ generate a QAbstractListModel with the proper roles in it.
 ```python
 from __future__ import annotations
 from PySide6.QtCore import QObject, Property, QCoreApplication
-from qtgql.itemsystem import get_base_type, role, GenericModel, asdict
+from qtgql.tools.itemsystem import get_base_type, role, GenericModel, asdict
 
 MyBaseType = get_base_type()
 class Apple(MyBaseType):
@@ -54,4 +54,4 @@ app.check_apple_color()
 ```
 
 
-::: itemsystem.core
+::: tools.itemsystem.core
