--- qtgql/itemsystem/__init__.py
@@ -1,3 +1,0 @@
-from .core import GenericModel, RoleDoesNotExist, asdict, get_base_type, role
-
-__all__ = ["role", "get_base_type", "GenericModel", "RoleDoesNotExist", "asdict"]
