+++ qtgql/tools/__init__.py
@@ -1,0 +1,5 @@
+from .qproperty import qproperty
+from .slot import slot
+from .autoproperty import define_properties
+
+__all__ = [slot, qproperty, define_properties]
