+++ qtgql/tools/autoproperty.py
