+++ qtgql/tools/itemsystem.py
@@ -8,14 +8,15 @@ import attrs
 from PySide6.QtCore import QAbstractListModel, QByteArray, QEventLoop, QModelIndex, Qt, Signal
 from typing_extensions import Self, dataclass_transform
 
-from qtgql import slot
+from qtgql.tools import slot
 from qtgql.exceptions import QtGqlException
-from qtgql.typingref import UNSET, TypeHinter
+from qtgql.utils.typingref import UNSET, TypeHinter
 
 IS_GQL = "is_gql"
 IS_ROLE = "is_role"
 OLD_NAME = "old_name"
 
+__all__ = ["role", "get_base_type", "GenericModel", "RoleDoesNotExist", "asdict"]
 
 def asdict(t: _BaseType) -> dict:
     return attrs.asdict(t)  # type: ignore
@@ -168,7 +169,7 @@ class GenericModel(Generic[_TBaseType], QAbstractListModel):
     Although you would need it to annotate "child" roles. i.e:
 
     ```python
-    from qtgql.itemsystem import get_base_type, GenericModel
+    from qtgql.tools.itemsystem import get_base_type, GenericModel
 
     BaseType = get_base_type()
     class SomeModel(BaseType):
