+++ tests/test_docs.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 import pytest
 from mktestdocs import check_docstring, check_md_file
-from qtgql.autoproperty import define_properties
+from qtgql.tools import define_properties
 from qtgql.codegen.py.runtime.custom_scalars import BaseCustomScalar
 
 docs_dir = Path(__file__).parent.parent / "docs"
