+++ tests/test_itemsystem/conftest.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 from uuid import uuid4
 
 import pytest
-from qtgql.itemsystem import GenericModel, get_base_type, role
-from qtgql.itemsystem.core import _BaseType
+from qtgql.tools.itemsystem import GenericModel, get_base_type, role
+from qtgql.tools.itemsystem import _BaseType
 
 NORMAL_GQL = "normal_gql"
 NORMAL_GQL_CAMELIZED = "normalGql"
