+++ tests/test_sample_ui/main.py
@@ -7,7 +7,7 @@ from typing import Optional
 from PySide6 import QtCore, QtGui, QtQml, QtQuick
 from PySide6.QtCore import QObject, Signal
 from PySide6.QtQml import QQmlApplicationEngine
-from qtgql import slot
+from qtgql.tools import slot
 from qtgql.codegen.py.compiler.config import QtGqlConfig
 from qtgql.gqltransport.client import GqlClientMessage, GqlWsTransportClient, HandlerProto
